+++ src/components/index.js
@@ -1,2 +1,3 @@
 export * from './Loc';
 export * from './make-localized-element';
+export * from './pseudolocalize';
