+++ src/extraction/index.js
@@ -3,4 +3,3 @@ export * from './constants';
 export * from './defaults';
 export * from './extraction';
 export * from './format-helper';
-export * from './pseudolocalize';
